place cli app into a different module, separated from the library module
