+++ .gitignore
@@ -26,7 +26,7 @@ hs_err_pid*
 # --- Gradle
 .gradle
 **/build/
-!src/**/build/
+!lib/src/**/build/
 
 # Ignore Gradle GUI config
 gradle-app.setting
