+++ build.gradle
@@ -15,36 +15,13 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.3.72'
     id 'io.gitlab.arturbosch.detekt' version "$detekt_ver"
     id 'org.jlleitschuh.gradle.ktlint' version "$ktlint_ver"
-    id 'com.github.johnrengelman.shadow' version "$shadowjar_ver"
 }
-
-apply plugin: 'application'
 
 group 'org.wycliffeassociates'
 version '1.0-SNAPSHOT'
 
 repositories {
     jcenter()
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    implementation "org.wycliffeassociates:kotlin-resource-container:0.6.0"
-    implementation "com.squareup.retrofit2:retrofit:$retrofit_ver"
-    implementation "org.slf4j:slf4j-api:$slf4j_api_ver"
-    implementation "org.slf4j:slf4j-simple:$slf4j_simple_ver"
-    implementation("com.github.ajalt.clikt:clikt:3.0.0")
-
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-
-    detektPlugins "io.gitlab.arturbosch.detekt:detekt-formatting:$detekt_ver"
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
 }
 
 detekt {
@@ -55,18 +32,4 @@ detekt {
 
 ktlint {
     ignoreFailures = true
-}
-
-application {
-    mainClassName = "org.wycliffeassociates.resourcecontainer.media.MainKt"
-}
-
-shadowJar {
-    archiveFileName = 'rcmediadownloader.jar'
-    mergeServiceFiles()
-    manifest {
-        attributes(
-                'Main-Class': "org.wycliffeassociates.resourcecontainer.media.MainKt"
-        )
-    }
 }
