+++ cliapp/src/main/kotlin/org/wycliffeassociates/resourcecontainer/media/cli/CIApplication.kt
@@ -1,13 +1,13 @@
-package org.wycliffeassociates.resourcecontainer.media
-
+package org.wycliffeassociates.resourcecontainer.media.cli
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
-import java.io.File
+import org.wycliffeassociates.resourcecontainer.media.RCMediaDownloader
 import org.wycliffeassociates.resourcecontainer.media.data.MediaDivision
 import org.wycliffeassociates.resourcecontainer.media.data.MediaType
 import org.wycliffeassociates.resourcecontainer.media.data.MediaUrlParameter
+import java.io.File
 
 class CIApplication : CliktCommand() {
     private val rcPath by option(
