+++ cliapp/src/main/kotlin/org/wycliffeassociates/resourcecontainer/media/cli/Main.kt
@@ -1,3 +1,3 @@
-package org.wycliffeassociates.resourcecontainer.media
+package org.wycliffeassociates.resourcecontainer.media.cli
 
 fun main(args: Array<String>) = CIApplication().main(args)
