+++ dependencies.gradle
@@ -5,5 +5,4 @@ ext {
     slf4j_api_ver = '2.0.0-alpha1'
     slf4j_simple_ver = '2.0.0-alpha1'
     retrofit_ver='2.9.0'
-    shadowjar_ver = '5.2.0'
 }
