+++ lib/build.gradle
@@ -1,0 +1,25 @@
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_ver"
+    }
+}
+
+plugins {
+    id 'kotlin'
+}
+
+repositories {
+    jcenter()
+}
+
+dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    implementation "org.wycliffeassociates:kotlin-resource-container:0.6.0"
+    implementation "com.squareup.retrofit2:retrofit:$retrofit_ver"
+    implementation "org.slf4j:slf4j-api:$slf4j_api_ver"
+    implementation "org.slf4j:slf4j-simple:$slf4j_simple_ver"
+    implementation("com.github.ajalt.clikt:clikt:3.0.0")
+}
