+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'RCMediaDownloader'
-
+include ':lib', ':cliapp'
